% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features_iterative.R
\name{transform_features_iterative}
\alias{transform_features_iterative}
\title{Transform Features (Iterative Logic)}
\usage{
transform_features_iterative(X, decisions)
}
\arguments{
\item{X}{Original predictor data frame.}

\item{decisions}{Output of \code{decide_variable_type_iterative}.}
}
\value{
A data frame with the chosen variables in their final forms (dummy
  or linear).
}
\description{
Once \code{decide_variable_type_iterative} has chosen which
  variables to add (and how), we can build a final data frame from those
  decisions.
}
\keyword{internal}
