% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features_univariate.R
\name{transform_features_univariate}
\alias{transform_features_univariate}
\title{Transform Features (Univariate Logic)}
\usage{
transform_features_univariate(X, decisions)
}
\arguments{
\item{X}{Original predictor data frame.}

\item{decisions}{The list returned by \code{decide_variable_type_univariate}.}
}
\value{
A new data frame with either the original column or a dummy column
  for each variable.
}
\description{
Given the decisions (dummy or linear) for each predictor,
  produce a transformed data frame. Dummy columns are 0/1 based on the
  cutoff.
}
\keyword{internal}
